+++ tests/ZohoDatabaseCopierTest.php
@@ -17,20 +17,6 @@ class ZohoDatabaseCopierTest extends \PHPUnit_Framework_TestCase
      */
     protected $dbConnection;
 
-    public static function setUpBeforeClass()
-    {
-        $config = new \Doctrine\DBAL\Configuration();
-        $connectionParams = array(
-            'user' => $GLOBALS['db_username'],
-            'password' => $GLOBALS['db_password'],
-            'host' => $GLOBALS['db_host'],
-            'port' => $GLOBALS['db_port'],
-            'driver' => $GLOBALS['db_driver'],
-        );
-        $adminConn = DriverManager::getConnection($connectionParams, $config);
-        $adminConn->getSchemaManager()->dropAndCreateDatabase($GLOBALS['db_name']);
-    }
-
     protected function setUp()
     {
         $config = new \Doctrine\DBAL\Configuration();
@@ -50,20 +36,12 @@ class ZohoDatabaseCopierTest extends \PHPUnit_Framework_TestCase
         return new ZohoClient($GLOBALS['auth_token']);
     }
 
-    public function getEntitiesGeneratorService()
-    {
-        return new EntitiesGeneratorService($this->getZohoClient(), new NullLogger());
-    }
-
-    public function testSync()
+    /**
+     * @depends Wabel\Zoho\CRM\Copy\ZohoDatabaseModelSyncTest::testModelSync
+     */
+    public function testFetch()
     {
         $listener = new TestListener();
-
-        $generator = $this->getEntitiesGeneratorService();
-        $generator->generateModule('Contacts', 'Contacts', 'Contact', __DIR__.'/generated/', 'TestNamespace');
-
-        require __DIR__.'/generated/Contact.php';
-        require __DIR__.'/generated/ContactZohoDao.php';
 
         $contactZohoDao = new ContactZohoDao($this->getZohoClient());
 
@@ -76,7 +54,7 @@ class ZohoDatabaseCopierTest extends \PHPUnit_Framework_TestCase
 
         $databaseCopier = new ZohoDatabaseCopier($this->dbConnection, 'zoho_', [$listener]);
 
-        $databaseCopier->copy($contactZohoDao);
+        $databaseCopier->fetchFromZoho($contactZohoDao);
 
         $this->assertTrue($listener->isInsertCalled());
 
@@ -89,7 +67,7 @@ class ZohoDatabaseCopierTest extends \PHPUnit_Framework_TestCase
         $testContact->setLastName(uniqid('Test2'));
         $contactZohoDao->save($testContact);
 
-        $databaseCopier->copy($contactZohoDao);
+        $databaseCopier->fetchFromZoho($contactZohoDao);
         $result = $this->dbConnection->fetchAssoc('SELECT * FROM zoho_contacts WHERE id = :id', ['id' => $testContact->getZohoId()]);
         $this->assertNotFalse($result);
         $this->assertEquals($testContact->getLastName(), $result['lastName']);
